[WIP] Middle of refactoring Camera access
Starting to get the Camera hooked up for video capture to grab images to receive code. Since
transmission also needs the Camera I think it will need to be shared and bound to the same
Composable. The Tx and Rx may be mutually exclusive, which would be unfortunate I guess...