[ImageAnalyzer] Add an average to try and make signals stand out
Lots more futzing with how I'm trying to capture signals. Ended with trying to
take a running average of the amount of "light" and then taking the difference
between the amount of light in a frame and the average. If it's above a certain
threshold it triggers a signal.

Next will continue working on a less clunky state machine.